+++ src/web/questions/pages/QuestionPage.tsx
@@ -6,10 +6,11 @@ import { Query } from "../../common/Query";
 import { Card } from "../../display/Card";
 import { DisplayOneQuestionForCapture } from "../../display/DisplayOneQuestionForCapture";
 import {
-    formatIndicatorValue, getStarsElement, qualityIndicatorLabels, UsedIndicatorName
+    formatIndicatorValue, qualityIndicatorLabels, UsedIndicatorName
 } from "../../display/DisplayQuestion/QuestionFooter";
 import { Layout } from "../../display/Layout";
 import { LineHeader } from "../../display/LineHeader";
+import { Stars } from "../../display/Stars";
 import { QuestionWithHistoryFragment } from "../../fragments.generated";
 import { ssrUrql } from "../../urql";
 import { HistoryChart } from "../components/HistoryChart";
@@ -44,7 +45,7 @@ export const getServerSideProps: GetServerSideProps<Props> = async (
 const QuestionCardContents: React.FC<{
   question: QuestionWithHistoryFragment;
 }> = ({ question }) => (
-  <div className="grid grid-cols-1 space-y-4 place-items-center mb-5">
+  <div className="grid grid-cols-1 space-y-4 place-items-center">
     <h1 className="text-4xl place-self-center w-full text-center mt-10 pl-5 pr-5">
       <a
         className="text-black no-underline hover:text-gray-600"
@@ -56,18 +57,9 @@ const QuestionCardContents: React.FC<{
       </a>
     </h1>
     <HistoryChart question={question} />
-    {/*
-        <div className="flex justify-center items-center w-full">
-      <div className="w-6/12">
-        <QuestionFooter question={question} expandFooterToFullWidth={true} />
-      </div>
-    </div>
-    <QuestionOptions options={question.options} />
-
-    */}
 
     <h2 className="pt-10 text-xl place-self-center w-full text-center text-gray-900">
-      {"Question description"}
+      Question description
     </h2>
     <ReactMarkdown
       linkTarget="_blank"
@@ -76,12 +68,12 @@ const QuestionCardContents: React.FC<{
       {question.description.replaceAll("---", "")}
     </ReactMarkdown>
 
-    <h2 className="pt-2  text-xl place-self-center w-full text-center text-gray-900">
-      {"Indicators"}
+    <h2 className="pt-2 text-xl place-self-center w-full text-center text-gray-900">
+      Indicators
     </h2>
     <div className="relative overflow-x-auto shadow-md sm:rounded-lg">
-      <table className="w-full text-sm text-left text-gray-500 dark:text-gray-400">
-        <thead className="text-xs text-gray-700 uppercase bg-gray-50 dark:bg-gray-700 dark:text-gray-400">
+      <table className="w-full text-sm text-left text-gray-500">
+        <thead className="text-xs text-gray-700 uppercase bg-gray-100">
           <tr>
             <th scope="col" className="px-6 py-3">
               Indicator
@@ -92,41 +84,39 @@ const QuestionCardContents: React.FC<{
           </tr>
         </thead>
         <tbody>
-          <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700">
+          <tr className="border-b">
             <th
               scope="row"
-              className="px-6 py-4 font-medium text-gray-900 dark:text-white whitespace-nowrap"
+              className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap"
             >
-              {"Stars"}
+              Stars
             </th>
             <td className="px-6 py-4">
-              {getStarsElement(question.qualityIndicators["stars"])}
+              <Stars num={question.qualityIndicators.stars} />
             </td>
           </tr>
-          <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700">
+          <tr className="border-b">
             <th
               scope="row"
-              className="px-6 py-4 font-medium text-gray-900 dark:text-white whitespace-nowrap"
+              className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap"
             >
-              {"Platform"}
+              Platform
             </th>
             <td className="px-6 py-4">{question.platform.label}</td>
           </tr>
-          {!!question.qualityIndicators["numForecasts"] ? (
-            <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700">
+          {question.qualityIndicators.numForecasts ? (
+            <tr className="border-b">
               <th
                 scope="row"
-                className="px-6 py-4 font-medium text-gray-900 dark:text-white whitespace-nowrap"
+                className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap"
               >
-                {"Number of forecasts"}
+                Number of forecasts
               </th>
               <td className="px-6 py-4">
-                {question.qualityIndicators["numForecasts"]}
+                {question.qualityIndicators.numForecasts}
               </td>
             </tr>
-          ) : (
-            ""
-          )}
+          ) : null}
           {Object.keys(question.qualityIndicators)
             .filter(
               (indicator) =>
@@ -135,10 +125,10 @@ const QuestionCardContents: React.FC<{
             )
             .map((indicator: UsedIndicatorName) => {
               return (
-                <tr className="bg-white border-b dark:bg-gray-800 dark:border-gray-700">
+                <tr className="bg-white border-b" key={indicator}>
                   <th
                     scope="row"
-                    className="px-6 py-4 font-medium text-gray-900 dark:text-white whitespace-nowrap"
+                    className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap"
                   >
                     {qualityIndicatorLabels[indicator]}
                   </th>
